+++ examples/relay/tasking/tasking_receive.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+$LOAD_PATH.unshift(File.dirname(__FILE__) + '/../lib')
+%w[
+  bundler/setup
+  signalwire
+].each { |f| require f }
+
+# Setup your ENV with:
+# SIGNALWIRE_ACCOUNT=YOUR_SIGNALWIRE_ACCOUNT_ID
+# SIGNALWIRE_TOKEN=YOUR_SIGNALWIRE_ACCOUNT_TOKEN
+#
+Signalwire::Logger.logger.level = ::Logger::DEBUG
+
+class MyConsumer < Signalwire::Relay::Consumer
+  contexts ['incoming']
+
+  def on_task(task)
+    logger.debug "Received #{task.message}"
+  end
+end
+
+MyConsumer.new.run