+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 ### Added
+- `:task` broadcast from client
+- `Relay::Task` and `on_task` handler for `Consumer`
 - Fax API for Relay
 
 ## [2.0.0] - 2019-07-16